TIL temp

+++ review/controllers/videoController.js
@@ -1,8 +1,8 @@
-// Video Controller
 import routes from "../routes";
 import Video from "../models/Video";
 
 // Home
+
 export const home = async (req, res) => {
   try {
     const videos = await Video.find({}).sort({ _id: -1 });
@@ -14,6 +14,7 @@ export const home = async (req, res) => {
 };
 
 // Search
+
 export const search = async (req, res) => {
   const {
     query: { term: searchingBy }
@@ -30,52 +31,54 @@ export const search = async (req, res) => {
 };
 
 // Upload
-export const getUpload = (req, res) => {
+
+export const getUpload = (req, res) =>
   res.render("upload", { pageTitle: "Upload" });
-};
 
 export const postUpload = async (req, res) => {
   const {
     body: { title, description },
     file: { path }
   } = req;
-  try {
-    const newVideo = await Video.create({
-      fileUrl: path,
-      title,
-      description
-    });
-    res.redirect(routes.videoDetail(newVideo.id));
-  } catch (error) {
-    console.log(error);
-    res.redirect(routes.upload);
-  }
+  const newVideo = await Video.create({
+    fileUrl: path,
+    title,
+    description,
+    creator: req.user.id
+  });
+  req.user.videos.push(newVideo.id);
+  req.user.save();
+  res.redirect(routes.videoDetail(newVideo.id));
 };
 
 // Video Detail
+
 export const videoDetail = async (req, res) => {
   const {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
+    const video = await Video.findById(id).populate("creator");
     res.render("videoDetail", { pageTitle: video.title, video });
   } catch (error) {
-    console.log(error);
     res.redirect(routes.home);
   }
 };
 
 // Edit Video
+
 export const getEditVideo = async (req, res) => {
   const {
     params: { id }
   } = req;
   try {
-    const video = await Video.findById(id);
-    res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    const video = await Video.findById(id).populate("video");
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      res.render("editVideo", { pageTitle: `Edit ${video.title}`, video });
+    }
   } catch (error) {
-    console.log(error);
     res.redirect(routes.home);
   }
 };
@@ -86,21 +89,26 @@ export const postEditVideo = async (req, res) => {
     body: { title, description }
   } = req;
   try {
-    await Video.findOneAndUpdate({ id }, { title, description });
+    await Video.findOneAndUpdate({ _id: id }, { title, description });
     res.redirect(routes.videoDetail(id));
   } catch (error) {
-    console.log(error);
     res.redirect(routes.home);
   }
 };
 
 // Delete Video
+
 export const deleteVideo = async (req, res) => {
   const {
     params: { id }
   } = req;
   try {
-    await Video.findOneAndRemove({ _id: id });
+    const video = await Video.findById(id);
+    if (video.creator !== req.user.id) {
+      throw Error();
+    } else {
+      await Video.findOneAndRemove({ _id: id });
+    }
   } catch (error) {
     console.log(error);
   }
